+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -283,7 +283,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-03-26T07:53:12.733Z",
+    "2025-03-26T14:07:36.344Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -328,7 +328,7 @@ export const UpdateDatasetResponseBody$outboundSchema: z.ZodType<
   parentId: z.string().optional(),
   version: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-03-26T07:53:12.733Z"))
+  updated: z.date().default(() => new Date("2025-03-26T14:07:36.344Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {